Make 'cmaes.cma' module private
